// src/types/types.ts
export interface User {
    average_rating: number | null
    bio: string
    clerkId: string
    createdAt: string
    email: string
    interests: string
    name: string
    ongoing_project_links: string[]
    phone_number: string | null
    portfolio_links: string[]
    role: string
    skills: string[]
    socials: string[]
    tags: string[]
    updatedAt: string
    verified: boolean
  }
  
  